requirements.txt
numpy
pandas
matplotlib
plotly[express]
seaborn
bokeh
streamlit
scipy
statsmodels
pingouin
pymc
mplcyberpunk
pyyaml
mysql-connector
python-dotenv
st_pages